refactor(acp): split sink.go into focused files
Separates the 450-line sink.go into five focused files:

- sink.go: Core struct, NewSink, Start, Stop
- sink_messages.go: HandleMessage, translatePart, translateText,
  translateReasoning
- sink_permissions.go: HandlePermission, extractEditParams
- sink_tools.go: translateToolCall, translateToolResult, parseToolInput,
  toolKind
- sink_plan.go: HandleSession (todo→plan conversion)
